virtualfit-enterprise-merged-with-tryon/services/inference-service/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
numpy
opencv-python
pydantic
python-multipart
